requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.11.3
pydantic-settings==2.8.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.22
alembic==1.12.1
pillow==10.0.1
redis==5.0.1
httpx==0.25.0
python-dotenv==1.0.0
boto3==1.28.64
pytest==7.4.3
email-validator==2.2.0
